Rename `flareFixtures.ts` to `serverlessFixtures.ts`
