+++ src/commands/lambda/__tests__/flare.test.ts
@@ -19,7 +19,7 @@ import {
   MOCK_CWD,
   MOCK_DATADOG_API_KEY,
   MOCK_FOLDER_PATH,
-} from '../../../helpers/__tests__/flareFixtures'
+} from '../../../helpers/__tests__/serverlessFixtures'
 import * as helpersPromptModule from '../../../helpers/prompt'
 
 import {PROJECT_FILES} from '../constants'
