+++ src/commands/lambda/__tests__/functions/instrument.part2.test.ts
@@ -1,7 +1,7 @@
 jest.mock('../../loggroup')
 
 import {CI_API_KEY_ENV_VAR} from '../../../../constants'
-import {MOCK_DATADOG_API_KEY} from '../../../../helpers/__tests__/flareFixtures'
+import {MOCK_DATADOG_API_KEY} from '../../../../helpers/__tests__/serverlessFixtures'
 
 import {CI_API_KEY_SECRET_ARN_ENV_VAR, CI_KMS_API_KEY_ENV_VAR} from '../../constants'
 import {calculateUpdateRequest} from '../../functions/instrument'
